+++ face-llm/src/test/java/fun/commons/tokengateway/controller/CountTokensControllerTest.java
@@ -53,7 +53,7 @@ class CountTokensControllerTest {
         var channelApi = new HttpChannelApi(b, new fun.commons.tokengateway.rpc.CapabilityEndpoints(new fun.commons.tokengateway.spi.config.TokenGatewayProperties(), props), new RpcInternalAuth(props));
         controller = new CountTokensController(
                 new RelayOrchestrator(tokenApi, channelApi, new HttpBillingApi(b, new fun.commons.tokengateway.rpc.CapabilityEndpoints(new fun.commons.tokengateway.spi.config.TokenGatewayProperties(), props), new RpcInternalAuth(props)),
-                        new ModerationGate(new fun.commons.tokengateway.rpc.HttpModerationApi(b, new fun.commons.tokengateway.rpc.CapabilityEndpoints(new fun.commons.tokengateway.spi.config.TokenGatewayProperties(), props), new RpcInternalAuth(props), new fun.commons.tokengateway.spi.config.TokenGatewayProperties())),
+                        new ModerationGate(new fun.commons.tokengateway.rpc.HttpModerationApi(b, new fun.commons.tokengateway.rpc.CapabilityEndpoints(new fun.commons.tokengateway.spi.config.TokenGatewayProperties(), props), new RpcInternalAuth(props), moderationOnSpi())),
                 new fun.commons.tokengateway.thmp.ThmpShadow.Noop(),
                 new fun.commons.tokengateway.thmp.ThmpCutover.Noop()),
                 new FormatConverter(), b,
@@ -65,6 +65,16 @@ class CountTokensControllerTest {
     void tearDown() throws Exception {
         backend.shutdown();
         upstream.shutdown();
+    }
+
+    /**
+     * issue #38: moderation.enabled 缺省 false 起 HttpModerationApi 双闸短路 (不发 RPC),
+     * 本夹具 mockDistribute 期待 scan RPC 真实下发, 显式开启.
+     */
+    private static fun.commons.tokengateway.spi.config.TokenGatewayProperties moderationOnSpi() {
+        var spi = new fun.commons.tokengateway.spi.config.TokenGatewayProperties();
+        spi.getModeration().setEnabled(true);
+        return spi;
     }
 
     private void mockTokenOk() {